Deleting the image files when deleting product

+++ controllers/admin.js
@@ -1,5 +1,7 @@
 const mongoose = require('mongoose');
 
+const fileHelper = require('../util/file');
+
 const {validationResult} = require('express-validator/check');
 
 const Product = require('../models/product');
@@ -162,6 +164,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDesc;
       if(image){
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save()
@@ -194,10 +197,27 @@ exports.getProducts = (req, res, next) => {
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
   // Product.findByIdAndRemove(prodId)
-    Product.deleteOne({_id: prodId, userId: req.user._id})
-    .then(() => {
-      console.log('DESTROYED PRODUCT');
-      res.redirect('/admin/products');
+    // Product.deleteOne({_id: prodId, userId: req.user._id})
+    // .then(() => {
+    //   console.log('DESTROYED PRODUCT');
+    //   res.redirect('/admin/products');
+    // })
+    // .catch(err => console.log(err));
+
+    //we will find the product and then delete the image also if deleting the product
+    Product.findById(prodId)
+    .then(product => {
+      if(!product){
+        return next(new Error('Product not Found!'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({_id: prodId, userId: req.user._id})
     })
-    .catch(err => console.log(err));
+    .then(() => {
+        console.log('DESTROYED PRODUCT');
+        res.redirect('/admin/products');
+      })
+    .catch(err => {
+      return next(new Error(err));
+    })
 };
